+++ src/main/java/com/watermelon/server/lottery/dto/response/ResponseLotteryWinnerInfoDto.java
@@ -1,6 +1,6 @@
 package com.watermelon.server.lottery.dto.response;
 
-import com.watermelon.server.lottery.domain.LotteryApplier;
+import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import lombok.Builder;
 import lombok.Data;
