+++ src/main/java/com/watermelon/server/lottery/parts/service/LotteryApplierPartsServiceImpl.java
@@ -1,6 +1,6 @@
 package com.watermelon.server.lottery.parts.service;
 
-import com.watermelon.server.lottery.domain.LotteryApplier;
+import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.lottery.parts.domain.LotteryApplierParts;
 import com.watermelon.server.lottery.parts.domain.Parts;
 import com.watermelon.server.lottery.parts.repository.LotteryApplierPartsRepository;
