+++ src/main/java/com/watermelon/server/lottery/repository/LotteryRewardRepository.java
@@ -1,6 +1,6 @@
 package com.watermelon.server.lottery.repository;
 
-import com.watermelon.server.lottery.domain.LotteryReward;
+import com.watermelon.server.event.lottery.domain.LotteryReward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
