+++ src/test/java/com/watermelon/server/lottery/link/service/LinkServiceImplTest.java
@@ -1,7 +1,7 @@
 package com.watermelon.server.lottery.link.service;
 
-import com.watermelon.server.lottery.domain.Link;
-import com.watermelon.server.lottery.domain.LotteryApplier;
+import com.watermelon.server.event.lottery.domain.Link;
+import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.lottery.link.dto.MyLinkDto;
 import com.watermelon.server.lottery.link.repository.LinkRepository;
 import com.watermelon.server.lottery.service.LotteryService;
