+++ src/test/java/com/watermelon/server/lottery/repository/ExpectationRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.watermelon.server.lottery.repository;
 
-import com.watermelon.server.lottery.domain.Expectation;
+import com.watermelon.server.event.lottery.domain.Expectation;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
